Check if cert files actually exist
Currently the code just checks if the variables have been set. If you followed the guide, this will be true. However, the cert files aren't actually at the location mentioned in the config file.

This one-line change checks if the files actually exist.

Optionally, the code could still be improved. If for some reason the user has specified a different cert files path than `./mycroft/personal_backend/certs/`, then this code is hard-wired to generate the files at that location anyway. It might be better to check which location the user has specified, and then try to generate the cert files at that location if they don't exist.